Imports miltiple STPs into one 3DM
This is a Python script that imports multiple ".stp" files into a Rhino 3D model. The script uses the RhinoScript and os libraries to import the ".stp" files. The user is prompted to enter the directory containing the ".stp" files, and the script then retrieves a list of all ".stp" files in the directory. The script loops through each ".stp" file in the list and imports it into the Rhino model by calling the rs.Command function. An exception handler is included to handle any errors that may occur when accessing the directory. If an error occurs, an error message is printed to the console indicating that there was an error accessing the directory.